Changed slogan and changed the text color from yellow to gold
Changed the header text of the "hero" section to be bigger and bold. Changed text color to blue.

Changed the subtext in the hero section to be smaller and removed the bold.

Added a black background overlay behind the text above the fold to improve visibility

Added margins between the elements in the hero section to make it more readable and feel more "open"